% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSpaces.R
\name{cleanSpaces}
\alias{cleanSpaces}
\title{Cleanup spaces on a text}
\usage{
cleanSpaces(text)
}
\arguments{
\item{text}{character cleaned.}
}
\value{
Returns a character without spaces at the begin and the end and replaces multiples spaces by one.
}
\description{
Remove the spaces of text, except the spaces between words.
}
\author{
Daniel Fischer
}
