% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanString.R
\name{cleanString}
\alias{cleanString}
\title{Clean characters on strings}
\usage{
cleanString(string, default_replacements = c(Á = "A", É = "E", Í =
  "I", Ó = "O", `Ú|Ü` = "U", Ñ = "N", á = "a", é = "e", í = "i", ó
  = "o", `ú|ü` = "u", ñ = "n"), extra_replacements = c(),
  allowed_chars = "a-zA-Z0-9", space_char = "_")
}
\arguments{
\item{string}{characters to be cleaned}

\item{default_replacements}{replacement patterns by default in format c("pattern" = "replacement")}

\item{extra_replacements}{extra replacements to be added to de default}

\item{allowed_chars}{in regular expression format, default "a-zA-Z0-9"}

\item{space_char}{character used for spaces, default _}
}
\description{
Clean characters on strings
}
\author{
Daniel Fischer
}
