% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAwsFilesRepo.R
\name{createAwsFilesRepo}
\alias{createAwsFilesRepo}
\title{Create folder for proyect files}
\usage{
createAwsFilesRepo(path = "files_project")
}
\arguments{
\item{repo_folder}{folder where repository is mounted}

\item{link_files}{should the repository be linked after beeing created?}
}
\description{
Create folder for proyect files
}
\author{
Daniel Fischer
}
