% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizePosix.R
\name{discretizePosix}
\alias{discretizePosix}
\title{Discretizes time periods}
\usage{
discretizePosix(timestamps, delta)
}
\arguments{
\item{timestamps}{array with timestamps to dscretize, must be POSIX.}

\item{delta}{time difference to discretize-with in seconds, use 60 to descretize by minutes,
3600 by hours, etc.}
}
\value{
Returns an array with times discretized.
}
\description{
Takes an array of timestamps and discretizes it by a specific time difference.
}
\author{
Daniel Fischer
}
