% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryptDelim.R
\name{encryptDelim}
\alias{encryptDelim}
\title{Encrypt Csv file}
\usage{
encryptDelim(input, output, encrypt_columns, algorithm = "sha256",
  chunk_size = 1e+05, ...)
}
\arguments{
\item{input}{input file}

\item{output}{output csv file}

\item{encrypt_columns}{columns to encrypt}

\item{algorithm}{encryption algorithm}

\item{chunk_size}{batch size}

\item{...}{arguments pass to read_delim.}
}
\description{
Encrypt Csv files
}
\author{
Daniel Fischer
}
