% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromBlob.R
\name{fromBlob}
\alias{fromBlob}
\title{Link Files folder}
\usage{
fromBlob(delete = F, account = "datarts", container = "proyects",
  path = "files_project", ...)
}
\description{
Move files from Blob
}
\author{
Daniel Fischer
}
