% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromS3.R
\name{fromS3}
\alias{fromS3}
\title{sync from S3}
\usage{
fromS3()
}
\arguments{
\item{delete}{shoud file be deleted localy?}

\item{path}{path to repository}

\item{...}{passed to system}

\item{bucket}{bucket name}
}
\value{
None
}
\description{
synchronizes from s3 proyect
}
\author{
Daniel Fischer
}
