% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDvRut.R
\name{getDvRut}
\alias{getDvRut}
\title{Calculates DV from RUT}
\usage{
getDvRut(rut, append_dv = F)
}
\arguments{
\item{rut}{vector of ruts on multiple format format (specified in examples)}

\item{append_dv}{should the calculated DV be apended to the rut or just return the DV}
}
\description{
Calculates DV from RUT
}
\examples{

ruts_sin_dv = c("15829332","16876895","17994104")
getDvRut(ruts_sin_dv)
getDvRut(ruts_sin_dv,append_dv = T)

ruts = "30.686.957-X"
getDvRut(ruts_sin_dv)
}
