% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLift.R
\name{getLift}
\alias{getLift}
\title{Provides Lift Curve}
\usage{
getLift(score, response, groups = 100)
}
\arguments{
\item{score}{a numeric vector with the scores from the prodictive model}

\item{response}{a boolean (TRUE or FALSE) vector with the real values from the dataset}

\item{groups}{number of percentiles, default 100}
}
\value{
data.frame with the values of the Lift curve.
score = runif(1000)
response = (score + rnorm(1000,0,0.1)) > 0.5
lift = getLift(score,response)
head(lift)
tail(lift)
}
\description{
Calculates the lift curve based on a score(numeric 0-1) and responce(TRUE/FALSE) vector
}
\author{
Daniel Fischer
}
