% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getROC.R
\name{getROC}
\alias{getROC}
\title{Calculates ROC Curve}
\usage{
getROC(score, response)
}
\arguments{
\item{score}{a numeric vector with the scores from the prodictive model}

\item{response}{a boolean (TRUE or FALSE) vector with the real values from the dataset}
}
\value{
dataframe with the values of the ROC curve.
}
\description{
Calculates the ROC curve
}
\examples{
score = runif(1000)
response = (score + rnorm(1000,0,0.1)) > 0.5
ROC = getROC(score, response)
plot(ROC$false_positive,ROC$true_positive,type="l")

}
