% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limitLevels.R
\name{limitLevels}
\alias{limitLevels}
\title{Truncates number of categories o a vector}
\usage{
limitLevels(vector, max_levels, other_name = "other")
}
\arguments{
\item{vector}{vector with character values}

\item{max_levels}{number of categories minuts the "other" category}

\item{category}{for the "other" category}
}
\description{
Limits the number of categoryes on a vector, it selects the most frecuent categories.
}
\examples{
vector = sample(letters[1:10],100,replace = T)
limitLevels(vector,5)

}
\author{
Daniel Fischer
}
