% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLift.R
\name{plotLift}
\alias{plotLift}
\title{Plots lifts from getLift function}
\usage{
plotLift(lifts = list(), ..., cuts = c(5, 15), value = "cum_lift")
}
\description{
Plots lifts ffrom getLift function
 @param lifts named list with lifts from getLift
 @param ... lifts from getLift
 @cuts points to show exact value of lifts
 @value column to be ploted, cum_lift, lift, cum_response, response
}
\examples{
score1 = runif(1000)
response1 = (score1 + rnorm(1000,0,0.1)) > 0.5
lift1 = getLift(score1,response1)
score2 = runif(1000)
response2 = (score2 + rnorm(1000,0,0.3)) > 0.5
lift2 = getLift(score2,response2)
lifts = list(model1 = lift1, model2 = lift2)
plotLift(lifts)
#or
plotLift(model1 = lift1, model2 = lift2) 

}
\author{
Daniel Fischer
}
