% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROC.R
\name{plotROC}
\alias{plotROC}
\title{plot ROC from getROC}
\usage{
plotROC(rocs = list(), ..., show_auc = T)
}
\description{
returns a ggplot object with the ROC for the provided models
}
\author{
Daniel Fischer
}
