% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_create_table.R
\name{rs_create_table}
\alias{rs_create_table}
\title{Create a table from scratch, guessing the table schema}
\usage{
rs_create_table(df, server_name, table_name = deparse(substitute(df)),
  split_files, wlm_slots = 1, sortkeys, sortkey_style = "compound",
  distkey, distkey_style = "even", compression = T)
}
\arguments{
\item{df}{a data frame}

\item{table_name}{the name of the table to create}

\item{wlm_slots}{amount of WLM slots to use for this bulk load http://docs.aws.amazon.com/redshift/latest/dg/tutorial-configuring-workload-management.html}

\item{sortkeys}{Column or columns to sort the table by}

\item{sortkey_style}{Sortkey style, can be compound or interleaved http://docs.aws.amazon.com/redshift/latest/dg/t_Sorting_data-compare-sort-styles.html}

\item{distkey}{Distkey column, can only be one, if chosen the table is distributed among clusters according to a hash of this column's value.}

\item{distkey_style}{Distkey style, can be even or all, for the key distribution use the distkey parameter. http://docs.aws.amazon.com/redshift/latest/dg/t_Distributing_data.html}

\item{compression}{Add encoding for columns whose compression algorithm is easy to guess, for the rest you should upload it to Redshift and run ANALYZE COMPRESSION}

\item{dbcon}{an RPostgres connection to the redshift server}
}
\description{
Create a table from scratch, guessing the table schema
}
\examples{
library(DBI)

a=data.frame(a=seq(1,10000), b=seq(10000,1))

\dontrun{
con <- dbConnect(RPostgres::Postgres(), dbname="dbname",
host='my-redshift-url.amazon.com', port='5439',
user='myuser', password='mypassword',sslmode='require')

rs_create_table(df=a, dbcon=con, table_name='testTable',
bucket="my-bucket", split_files=4)

}
}
