% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeLibrary.R
\name{safeLibrary}
\alias{safeLibrary}
\title{installs non existing libraries}
\usage{
safeLibrary(x)
}
\arguments{
\item{x}{library to use}

\item{load}{load packages after checking if is installed}
}
\description{
It loads libraries, if the library is not installed it would install it first.
}
\author{
Daniel Fischer
}
