% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDataset.R
\name{sampleDataset}
\alias{sampleDataset}
\title{Samples dataframe}
\usage{
sampleDataset(dataset, sampleSize)
}
\arguments{
\item{dataset}{dataset to be sampled}

\item{sampleSize}{size of the sample

 @export}
}
\description{
Extract a mber of rows randomly from a fatafram
}
