% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLevels.R
\name{setLevels}
\alias{setLevels}
\title{Sets the categories of a dataset given a sample}
\usage{
setLevels(vector, level_values, other_name = "other")
}
\arguments{
\item{vector}{vector with character values}

\item{level_values}{sample with avivable categories}

\item{category}{for the "other" category}
}
\description{
Any category not included in level_values, will be classified as other_name, 
if the provided level_values is a factor type, 
it will use the levles(level_values) as the list of categoryes
}
\examples{
vector = sample(letters[1:10],100,replace = T)
limitLevels(vector,c("a","b","b"))

}
\author{
Daniel Fischer
}
