% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setResponseRate.R
\name{setResponseRate}
\alias{setResponseRate}
\title{Sets response rate for dataset}
\usage{
setResponseRate(dataset, response_variable, ...)
}
\arguments{
\item{dataset}{input dataframe}

\item{response_variable}{column to be resampled}

\item{...}{porportional number of cases per value of response_variable column}
}
\description{
Sets response rate for dataset
}
\examples{
setResponseRate(mtcars,"cyl", "4" = 4, "6" = 1, "8" = 1)

}
\author{
Daniel Fischer
}
