% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toS3.R
\name{toS3}
\alias{toS3}
\title{sync to S3}
\usage{
toS3()
}
\arguments{
\item{delete}{shoud file be deleted on s3?}

\item{path}{path to repository}

\item{bucket}{bucket name}
}
\value{
None
}
\description{
synchronizes to s3 proyect
}
\author{
Daniel Fischer
}
